feat(seo): server-render the community funding-opportunities directory
The funding-opportunities page rendered ~406 chars of client shell without
JavaScript, so the literal answer page for funder discovery was invisible to
crawlers. Prefetch the community's program list server-side (React.cache +
HydrationBoundary, one indexer round-trip) and hand it to the client tree as a
hydrated React Query entry, so every opportunity's title, status, deadline and
funding facts ship in the initial HTML.

Now that the content is real, flip the listing decisions from DEV-586 back:

- declare a whitelabel-aware self-referential canonical via
  communitySubpageMetadata
- re-add the route to the communities sitemap (priority 0.8 under the root)
- retire the funding-opportunities entry from the crawler's
  DEFAULT_KNOWN_ISSUES: a header-only page is a regression again, not an
  excused known issue

An indexer failure during prefetch degrades to the previous behaviour: the
skeleton ships and the client refetches — an outage is never hydrated as an
empty directory.
