+++ app/community/[communityId]/(with-header)/funding-opportunities/layout.tsx
@@ -1,5 +1,6 @@
 import type { Metadata } from "next";
 import { PROJECT_NAME } from "@/constants/brand";
+import { communitySubpageMetadata } from "@/utilities/metadata/communityCanonical";
 import { getCommunityDetails } from "@/utilities/queries/v2/getCommunityData";
 
 type LayoutProps = {
@@ -7,9 +8,10 @@ type LayoutProps = {
   params: Promise<{ communityId: string }>;
 };
 
-// The page itself is a client component, so its metadata lives here. Without it
-// the funding directory inherits the community layout's canonical and title,
-// which made a sitemap-listed URL point at `/community/<id>`.
+// The rendered directory lives in a client component, so its metadata lives
+// here. Without it the funding directory inherits the community layout's
+// canonical and title, which made a sitemap-listed URL point at
+// `/community/<id>`.
 export async function generateMetadata({
   params,
 }: {
@@ -19,11 +21,14 @@ export async function generateMetadata({
   const community = await getCommunityDetails(communityId);
   const communityName = community?.details?.name || communityId;
 
-  // No self-canonical: this route is a client-rendered shell, so it is absent
-  // from the sitemap and consolidates onto the community root canonical it
-  // inherits from the layout. Give it server-rendered content first, then a
-  // canonical and a sitemap entry together.
+  // Self-canonical (whitelabel-aware): the page now server-renders the program
+  // directory into the initial HTML, so it carries its own crawlable content
+  // and ships in the communities sitemap again — canonical, content and
+  // sitemap entry moved together (DEV-611).
+  const canonicalMetadata = await communitySubpageMetadata(communityId, "funding-opportunities");
+
   return {
+    ...canonicalMetadata,
     title: `${communityName} Funding Opportunities | ${PROJECT_NAME}`,
     description: `Find open, upcoming, and closed grant programs from ${communityName}. Compare funding amounts, deadlines, and eligibility before you apply on ${PROJECT_NAME}.`,
   };
