+++ src/features/programs/hooks/use-programs.ts
@@ -1,9 +1,26 @@
 import { useQuery } from "@tanstack/react-query";
 import { useEffect, useMemo } from "react";
+import { wlQueryKeys } from "@/src/lib/query-keys";
 import type { FundingProgram, ProgramFilters, ProgramStatus } from "@/types/whitelabel-entities";
 import { api } from "@/utilities/api/client";
+import { INDEXER } from "@/utilities/indexer";
 import { useProgramsStore } from "../lib/store";
 import type { UseProgramsReturn } from "../types";
+
+/**
+ * Shared with the server prefetch in `funding-opportunities/page.tsx`. The
+ * hydrated cache entry is only considered fresh — and therefore not refetched
+ * on mount — while both sides agree on this window.
+ */
+export const PROGRAMS_LIST_STALE_TIME = 5 * 60 * 1000;
+
+/**
+ * Default page size when no explicit limit filter is set. The server prefetch
+ * has no filter store, so it hydrates with this value — the two sides must
+ * agree or the hydrated entry would be structurally different from a client
+ * fetch.
+ */
+export const DEFAULT_PROGRAMS_LIMIT = 20;
 
 function matchesStatus(program: FundingProgram, status: ProgramStatus): boolean {
   const now = new Date();
@@ -46,16 +63,16 @@ export function usePrograms(
     programs: FundingProgram[];
     limit: number;
   }>({
-    queryKey: ["wl-programs", communityId],
+    queryKey: wlQueryKeys.programs.communityList(communityId),
     queryFn: async () => {
-      const limit = filters.limit || 20;
+      const limit = filters.limit || DEFAULT_PROGRAMS_LIMIT;
       // TODO(#1775): add zod schema
       const res = await api.get<FundingProgram[]>(
-        `/v2/funding-program-configs/community/${communityId}`
+        INDEXER.V2.FUNDING_PROGRAMS.BY_COMMUNITY(encodeURIComponent(communityId))
       );
       return { programs: res ?? [], limit };
     },
-    staleTime: 5 * 60 * 1000,
+    staleTime: PROGRAMS_LIST_STALE_TIME,
     enabled: !!communityId,
   });
